Fixed #8 patternomaly could not be included
The `dist` folder was not included in the npm package, also it was not possible to import/require the module as the documentation stated.

The `dist` folder has now been added to the Git ignore list and will now be included in NPM thanks to the presence of an .npmignore file.

Also updated the package.json file to include a browser property.